Factor out the fair islands generator
This breaks up the very large mapgen.cpp, making it more readable.
