+++ server/generator/fair_islands.h
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: GPLv3-or-later
+// SPDX-FileCopyrightText: Freeciv21 authors
+// SPDX-FileCopyrightText: Freeciv authors
+// SPDX-FileCopyrightText: Louis Moureaux <m_louis30@yahoo.com>
+
+bool map_generate_fair_islands();
